requirements.txt
# PyTorch with CUDA 12.1 support
# Install first: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# ML & data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# HuggingFace
huggingface_hub>=0.20.0
datasets>=2.14.0

# LangGraph pipeline
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM providers
groq>=0.9.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
shap>=0.43.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
